Improving native GraalOS termination (lambda creation after shutdown still not working).

+++ lambda-manager/src/scripts/stop_graalos_native.sh
@@ -23,8 +23,22 @@ curl -s http://localhost:$GRAALOS_PORT/exit
 
 LAMBDA_PID=$(cat "$LAMBDA_HOME"/lambda.pid)
 
-#while kill -0 $LAMBDA_PID; do
-#    sleep 0.5
-#done
+declare -i attempts=0
+terminated=""
+while kill -0 $LAMBDA_PID; do
+    attempts=$((attempts+1))
+    echo "Attempt #$attempts."
+    if [ "$attempts" -gt "100" ]; then
+        # Giving up on waiting for a graceful shutdown.
+        terminated="false"
+        break
+    fi
 
-kill $LAMBDA_PID
+    sleep 0.5
+done
+
+# Killing if couldn't shut down gracefully.
+if [[ "$terminated" == "false" ]]; then
+    echo "Killing the lambda manually."
+    kill $LAMBDA_PID
+fi
